// src/lib/gen/carbon/v1/landscape_pb.ts
// @generated by protoc-gen-es v2.14.0 with parameter "target=ts"
// @generated from file carbon/v1/landscape.proto (package carbon.v1, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { GraphEdge, GraphNode } from "./graph_pb";
import { file_carbon_v1_graph } from "./graph_pb";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file carbon/v1/landscape.proto.
 */
export const file_carbon_v1_landscape: GenFile = /*@__PURE__*/
  fileDesc("ChljYXJib24vdjEvbGFuZHNjYXBlLnByb3RvEgljYXJib24udjEiWgoOTGFuZHNjYXBlR3JhcGgSIwoFbm9kZXMYASADKAsyFC5jYXJib24udjEuR3JhcGhOb2RlEiMKBWVkZ2VzGAIgAygLMhQuY2FyYm9uLnYxLkdyYXBoRWRnZWIGcHJvdG8z", [file_carbon_v1_graph]);

/**
 * @generated from message carbon.v1.LandscapeGraph
 */
export type LandscapeGraph = Message<"carbon.v1.LandscapeGraph"> & {
  /**
   * @generated from field: repeated carbon.v1.GraphNode nodes = 1;
   */
  nodes: GraphNode[];

  /**
   * @generated from field: repeated carbon.v1.GraphEdge edges = 2;
   */
  edges: GraphEdge[];
};

/**
 * Describes the message carbon.v1.LandscapeGraph.
 * Use `create(LandscapeGraphSchema)` to create a new message.
 */
export const LandscapeGraphSchema: GenMessage<LandscapeGraph> = /*@__PURE__*/
  messageDesc(file_carbon_v1_landscape, 0);